+++ package.json
@@ -7,7 +7,7 @@
     "bottom-tip": "^0.1.5",
     "query-string": "^8.1.0",
     "typescript": "^5.3.2",
-    "vite": "^5.0.4"
+    "vite": "^5.0.5"
   },
   "dependencies": {
     "@calcit/procs": "^0.8.14",
