+++ src/apps/attractor-aizawa.mts
@@ -1,6 +1,6 @@
 import { createRenderer } from "../index.mjs";
-import attractorSprite from "../../shaders/attractor-sprite.wgsl?raw";
-import attractorCompute from "../../shaders/attractor-compute-aizawa.wgsl?raw";
+import attractorSprite from "./attractor/sprites.wgsl?raw";
+import attractorCompute from "./attractor/compute-aizawa.wgsl?raw";
 import { fiboGridN, rand_middle } from "../math.mjs";
 
 export let loadAizawaRenderer = async (canvas: HTMLCanvasElement) => {
