+++ src/apps/attractor/compute-aizawa.wgsl
