+++ src/apps/attractor/compute-four-wing.wgsl
