+++ src/apps/attractor/compute-lorenz.wgsl
