+++ src/apps/attractor/sprites.wgsl
