+++ src/apps/ball-spin.mts
@@ -1,6 +1,6 @@
 import { createRenderer } from "../index.mjs";
-import spriteShader from "../../shaders/ball-spin-sprites.wgsl?raw";
-import computeShader from "../../shaders/ball-spin-compute.wgsl?raw";
+import spriteShader from "./ball-spin/sprites.wgsl?raw";
+import computeShader from ".//ball-spin/compute.wgsl?raw";
 import { fiboGridN } from "../math.mjs";
 
 export let loadBallSpinRenderer = async (canvas: HTMLCanvasElement) => {
