+++ src/apps/ball-spin/compute.wgsl
