+++ src/apps/ball-spin/sprites.wgsl
