+++ src/apps/bifurcation.mts
@@ -1,6 +1,6 @@
 import { createRenderer } from "../index.mjs";
-import spriteShader from "../../shaders/bifurcation-sprites.wgsl?raw";
-import computeShader from "../../shaders/bifurcation-compute.wgsl?raw";
+import spriteShader from "./bifurcation/sprites.wgsl?raw";
+import computeShader from "./bifurcation/compute.wgsl?raw";
 import { rand_middle } from "../math.mjs";
 
 export let loadBifurcationRenderer = async (canvas: HTMLCanvasElement) => {
