+++ src/apps/bifurcation/compute.wgsl
