+++ src/apps/bifurcation/sprites.wgsl
