+++ src/apps/bounce.mts
@@ -1,6 +1,6 @@
 import { createRenderer } from "../index.mjs";
-import spriteShader from "../../shaders/bounce-sprites.wgsl?raw";
-import computeShader from "../../shaders/bounce-compute.wgsl?raw";
+import spriteShader from "./bounce/sprites.wgsl?raw";
+import computeShader from "./bounce/compute.wgsl?raw";
 import { fiboGridN } from "../math.mjs";
 
 export let loadBounceRenderer = async (canvas: HTMLCanvasElement) => {
