+++ src/apps/bounce/sprites.wgsl
