+++ src/apps/collision.mts
@@ -1,6 +1,6 @@
 import { createRenderer } from "../index.mjs";
-import spriteWGSL from "../../shaders/collision-sprites.wgsl?raw";
-import computeCollision from "../../shaders/collision-compute.wgsl?raw";
+import spriteWGSL from "./collision/sprites.wgsl?raw";
+import computeCollision from "./collision/compute.wgsl?raw";
 import { fiboGridN } from "../math.mjs";
 
 export let loadCollisionRenderer = async (canvas: HTMLCanvasElement) => {
