+++ src/apps/collision/sprites.wgsl
