+++ src/apps/fireworks/compute.wgsl
