+++ src/apps/fireworks/sprites.wgsl
