+++ src/apps/fractal/compute.wgsl
