+++ src/apps/fractal/sprites.wgsl
