+++ src/apps/lifegame/compute.wgsl
