+++ src/apps/lifegame/sprites.wgsl
